+++ src/main/java/lotto/LottoController.java
@@ -4,9 +4,9 @@ import lotto.domain.PaymentHistory;
 import lotto.domain.PaymentReceipt;
 import lotto.domain.Quantity;
 import lotto.domain.product.FinalResult;
-import lotto.domain.product.LotteryTicket;
-import lotto.domain.product.LotteryTickets;
-import lotto.domain.product.WinningTicket;
+import lotto.domain.product.lotto.LotteryTicket;
+import lotto.domain.product.lotto.LotteryTickets;
+import lotto.domain.product.lotto.WinningTicket;
 import lotto.view.input.ConsoleInputView;
 import lotto.view.input.InputView;
 import lotto.view.output.OutputView;
