+++ src/main/java/lotto/domain/product/lotto/LottoRank.java
@@ -1,4 +1,4 @@
-package lotto.domain.product;
+package lotto.domain.product.lotto;
 
 import java.util.Arrays;
