+++ src/main/java/lotto/view/input/InputView.java
@@ -2,9 +2,9 @@ package lotto.view.input;
 
 import lotto.domain.PaymentReceipt;
 import lotto.domain.Quantity;
-import lotto.domain.product.LotteryTickets;
-import lotto.domain.product.LottoNumber;
-import lotto.domain.product.WinningTicket;
+import lotto.domain.product.lotto.LotteryTickets;
+import lotto.domain.product.lotto.LottoNumber;
+import lotto.domain.product.lotto.WinningTicket;
 
 public interface InputView {
