+++ src/test/java/lotto/view/InputViewTest.java
@@ -2,10 +2,10 @@ package lotto.view;
 
 import lotto.domain.PaymentReceipt;
 import lotto.domain.Quantity;
-import lotto.domain.product.LotteryTicket;
-import lotto.domain.product.LottoNumber;
-import lotto.domain.product.LottoRank;
-import lotto.domain.product.WinningTicket;
+import lotto.domain.product.lotto.LotteryTicket;
+import lotto.domain.product.lotto.LottoNumber;
+import lotto.domain.product.lotto.LottoRank;
+import lotto.domain.product.lotto.WinningTicket;
 import lotto.view.fake.FakeInputView;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
