+++ src/test/java/lotto/view/fake/FakeInputView.java
@@ -3,10 +3,10 @@ package lotto.view.fake;
 
 import lotto.domain.PaymentReceipt;
 import lotto.domain.Quantity;
-import lotto.domain.product.LotteryTicket;
-import lotto.domain.product.LotteryTickets;
-import lotto.domain.product.LottoNumber;
-import lotto.domain.product.WinningTicket;
+import lotto.domain.product.lotto.LotteryTicket;
+import lotto.domain.product.lotto.LotteryTickets;
+import lotto.domain.product.lotto.LottoNumber;
+import lotto.domain.product.lotto.WinningTicket;
 import lotto.view.input.BaseInputView;
 
 public class FakeInputView extends BaseInputView {
